+++ ui/litellm-dashboard/src/components/usage/components/EntityUsage/entity_usage.tsx
@@ -21,21 +21,21 @@ import {
   TabPanels,
   Subtitle,
 } from "@tremor/react";
-import { ActivityMetrics, processActivityData } from "./activity_metrics";
-import { DailyData, BreakdownMetrics, KeyMetricWithMetadata, EntityMetricWithMetadata, TagUsage } from "./usage/types";
+import { ActivityMetrics, processActivityData } from "../../../activity_metrics";
+import { DailyData, BreakdownMetrics, KeyMetricWithMetadata, EntityMetricWithMetadata, TagUsage } from "../../types";
 import {
   organizationDailyActivityCall,
   tagDailyActivityCall,
   teamDailyActivityCall,
   customerDailyActivityCall,
   agentDailyActivityCall,
-} from "./networking";
+} from "../../../networking";
 import TopKeyView from "./top_key_view";
 import { formatNumberWithCommas } from "@/utils/dataUtils";
-import { valueFormatterSpend } from "./usage/utils/value_formatters";
-import { getProviderLogoAndName } from "./provider_info_helpers";
-import { UsageExportHeader } from "./EntityUsageExport";
-import type { EntityType } from "./EntityUsageExport/types";
+import { valueFormatterSpend } from "../../utils/value_formatters";
+import { getProviderLogoAndName } from "../../../provider_info_helpers";
+import { UsageExportHeader } from "../../../EntityUsageExport";
+import type { EntityType } from "../../../EntityUsageExport/types";
 import TopModelView from "./top_model_view";
 
 interface EntityMetrics {
