+++ ui/litellm-dashboard/src/components/usage/components/new_usage.tsx
@@ -33,22 +33,22 @@ import React, { useCallback, useEffect, useMemo, useState } from "react";
 import { useCustomers } from "@/app/(dashboard)/hooks/customers/useCustomers";
 import { formatNumberWithCommas } from "@/utils/dataUtils";
 import { Button } from "@tremor/react";
-import { all_admin_roles } from "../utils/roles";
-import { ActivityMetrics, processActivityData } from "./activity_metrics";
-import CloudZeroExportModal from "./cloudzero_export_modal";
-import EntityUsage, { EntityList } from "./entity_usage";
-import EntityUsageExportModal from "./EntityUsageExport";
-import { Team } from "./key_team_helpers/key_list";
-import { Organization, tagListCall, userDailyActivityAggregatedCall, userDailyActivityCall } from "./networking";
-import { getProviderLogoAndName } from "./provider_info_helpers";
-import AdvancedDatePicker from "./shared/advanced_date_picker";
-import { ChartLoader } from "./shared/chart_loader";
-import { Tag } from "./tag_management/types";
-import TopKeyView from "./top_key_view";
-import { DailyData, KeyMetricWithMetadata, MetricWithMetadata } from "./usage/types";
-import { valueFormatterSpend } from "./usage/utils/value_formatters";
-import UserAgentActivity from "./user_agent_activity";
-import ViewUserSpend from "./view_user_spend";
+import { all_admin_roles } from "../../../utils/roles";
+import { ActivityMetrics, processActivityData } from "../../activity_metrics";
+import CloudZeroExportModal from "../../cloudzero_export_modal";
+import EntityUsage, { EntityList } from "./EntityUsage/entity_usage";
+import EntityUsageExportModal from "../../EntityUsageExport";
+import { Team } from "../../key_team_helpers/key_list";
+import { Organization, tagListCall, userDailyActivityAggregatedCall, userDailyActivityCall } from "../../networking";
+import { getProviderLogoAndName } from "../../provider_info_helpers";
+import AdvancedDatePicker from "../../shared/advanced_date_picker";
+import { ChartLoader } from "../../shared/chart_loader";
+import { Tag } from "../../tag_management/types";
+import TopKeyView from "./EntityUsage/top_key_view";
+import { DailyData, KeyMetricWithMetadata, MetricWithMetadata } from "../types";
+import { valueFormatterSpend } from "../utils/value_formatters";
+import UserAgentActivity from "../../user_agent_activity";
+import ViewUserSpend from "../../view_user_spend";
 import { useAgents } from "@/app/(dashboard)/hooks/agents/useAgents";
 
 interface NewUsagePageProps {
